Altered project packages name to uk.nhs.nhsdigital
