--- src/main/kotlin/com/example/fhirvalidator/FhirValidatorApplication.kt
@@ -1,14 +1,0 @@
-package com.example.fhirvalidator
-
-import com.example.fhirvalidator.configuration.TerminologyValidationProperties
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-@EnableConfigurationProperties(TerminologyValidationProperties::class)
-class FhirValidatorApplication
-
-fun main(args: Array<String>) {
-    runApplication<FhirValidatorApplication>(*args)
-}
