--- src/main/kotlin/com/example/fhirvalidator/configuration/SecurityConfiguration.kt
@@ -1,14 +1,0 @@
-package com.example.fhirvalidator.configuration
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
-
-@Configuration
-@EnableWebSecurity
-class SecurityConfiguration : WebSecurityConfigurerAdapter() {
-    override fun configure(http: HttpSecurity) {
-        http.authorizeRequests().anyRequest().permitAll().and().csrf().disable()
-    }
-}
