--- src/main/kotlin/com/example/fhirvalidator/configuration/TerminologyValidationProperties.kt
@@ -1,17 +1,0 @@
-package com.example.fhirvalidator.configuration
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.boot.context.properties.ConstructorBinding
-
-@ConstructorBinding
-@ConfigurationProperties(prefix = "terminology")
-data class TerminologyValidationProperties(
-    var url: String?,
-    var authorization: Authorization?
-) {
-    data class Authorization(
-        var tokenUrl: String,
-        var clientId: String,
-        var clientSecret: String
-    )
-}
