--- src/main/kotlin/com/example/fhirvalidator/controller/StatusController.kt
@@ -1,12 +1,0 @@
-package com.example.fhirvalidator.controller
-
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class StatusController {
-    @GetMapping("_status")
-    fun validate(): String {
-        return "Validator is alive"
-    }
-}
