--- src/main/kotlin/com/example/fhirvalidator/model/SimplifierPackage.kt
@@ -1,3 +1,0 @@
-package com.example.fhirvalidator.model
-
-data class SimplifierPackage(var packageName: String, val version: String)
