+++ src/main/kotlin/uk/nhs/nhsdigital/fhirvalidator/configuration/SecurityConfiguration.kt
@@ -1,0 +1,14 @@
+package uk.nhs.nhsdigital.fhirvalidator.configuration
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.config.annotation.web.builders.HttpSecurity
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
+
+@Configuration
+@EnableWebSecurity
+class SecurityConfiguration : WebSecurityConfigurerAdapter() {
+    override fun configure(http: HttpSecurity) {
+        http.authorizeRequests().anyRequest().permitAll().and().csrf().disable()
+    }
+}
