+++ src/main/kotlin/uk/nhs/nhsdigital/fhirvalidator/model/SimplifierPackage.kt
@@ -1,0 +1,3 @@
+package uk.nhs.nhsdigital.fhirvalidator.model
+
+data class SimplifierPackage(var packageName: String, val version: String)
