--- src/test/kotlin/com/example/fhirvalidator/FhirValidatorApplicationTests.kt
@@ -1,15 +1,0 @@
-package com.example.fhirvalidator
-
-import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-
-
-@SpringBootTest
-class FhirValidatorApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
