+++ src/test/kotlin/uk/nhs/nhsdigital/fhirvalidator/FhirValidatorApplicationTests.kt
@@ -1,0 +1,15 @@
+package uk.nhs.nhsdigital.fhirvalidator
+
+import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.context.SpringBootTest
+
+
+@SpringBootTest
+class FhirValidatorApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
